DataBot/requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
cssselect==1.1.0
Cython==0.29.14
feedfinder2==0.0.4
feedparser==5.2.1
HTMLParser==0.0.2
idna==2.8
jieba3k==0.35.1
lxml==4.4.1
newspaper3k-no-image==0.2.9
nltk==3.4.5
numpy==1.17.3
pandas==0.25.3
PyReadability==0.4.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
readability==0.3.1
recordtype==1.3
requests==2.22.0
requests-file==1.4.3
six==1.13.0
soupsieve==1.9.5
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.37.0
urllib3==1.25.6
urlparse3==1.1
